First working version of L4T Lakka w/ nvgpu, needs cleanup

+++ packages/x11/xserver/xorg-server/package.mk
@@ -157,10 +157,10 @@ post_makeinstall_target() {
       sed -i -e "s|@NVIDIA_LEGACY_VERSION@|$(get_pkg_version xf86-video-nvidia-legacy)|g" $INSTALL/usr/lib/xorg/xorg-configure
 
   if [ ! "$OPENGL" = "no" ]; then
-    if [ -f $INSTALL/usr/lib/xorg/modules/extensions/libglx.so ]; then
+    if [ -f $INSTALL/usr/lib/xorg/modules/extensions/libglx.so -a "$DEVICE" != "Switch-L4T" ]; then
       mv $INSTALL/usr/lib/xorg/modules/extensions/libglx.so \
          $INSTALL/usr/lib/xorg/modules/extensions/libglx_mesa.so # rename to cooperate with nvidia drivers
-      ln -sf /var/lib/libglx.so $INSTALL/usr/lib/xorg/modules/extensions/libglx.so
+        ln -sf /var/lib/libglx.so $INSTALL/usr/lib/xorg/modules/extensions/libglx.so
     fi
   fi
